Prevent left nav overlapping content during scroll
When viewed on a smaller screen the content can be scrolled not only up
and down, but also left and right. More specifically it can be scrolled
back over the navigation menu to the left.

This prevents that, and as an bonus it fixes the scrolling to the right
where there is no content.